IMS-21 # Item controller completed

+++ src/main/java/com/qa/ims/persistence/dao/ItemDAO.java
@@ -22,7 +22,7 @@ public class ItemDAO implements Dao<Item> {
 	public Item modelFromResultSet(ResultSet resultSet) throws SQLException {
 		Long itemId = resultSet.getLong("itemId");
 		String itemName = resultSet.getString("itemName");
-		float price = resultSet.getFloat("price");
+		Double price = resultSet.getDouble("price");
 		return new Item(itemId, itemName, price);
 	}
 	
@@ -79,7 +79,7 @@ public class ItemDAO implements Dao<Item> {
 				PreparedStatement statement = connection
 						.prepareStatement("INSERT INTO Item(itemName, price) VALUES (?, ?)");) {
 			statement.setString(1, t.getItemName());
-			statement.setFloat(2, t.getPrice());
+			statement.setDouble(2, t.getPrice());
 			statement.executeUpdate();
 			return readLatest();
 		} catch (Exception e) {
@@ -95,7 +95,7 @@ public class ItemDAO implements Dao<Item> {
 				PreparedStatement statement = connection
 						.prepareStatement("UPDATE Item SET itemName = ?, price = ? WHERE itemId = ?");) {
 			statement.setString(1, t.getItemName());
-			statement.setFloat(2, t.getPrice());
+			statement.setDouble(2, t.getPrice());
 			statement.setLong(3, t.getItemId());
 			statement.executeUpdate();
 			return read(t.getItemId());
